+++ apps/shop/src/features/product/components/BannerSlide.module.scss
@@ -1,0 +1,9 @@
+.swiper {
+  width: 100%;
+  height: 100%;
+}
+
+.swiper-slide {
+  width: 100%;
+  height: 100%;
+}
