+++ apps/shop/src/features/product/components/ProductList.module.scss
@@ -1,10 +1,3 @@
-.list {
-  @include flex-wrap;
-  justify-content: space-between;
-  gap: 16px;
-  padding-bottom: 32px;
-}
-
 .loading {
   @include flex-center;
   padding: 2rem 0;
@@ -16,20 +9,4 @@
 
 .banner {
   padding-bottom: 32px;
-}
-
-.swiper {
-  width: 100%;
-  height: 100%;
-}
-
-.swiper-slide {
-  width: 100%;
-  height: 100%;
-}
-
-.title {
-  font-weight: 700;
-  line-height: 1.5rem;
-  padding-bottom: 16px;
 }
