+++ apps/shop/src/features/product/components/ProductSection.module.scss
@@ -1,0 +1,11 @@
+.title {
+  font-weight: 700;
+  line-height: 1.5rem;
+  padding-bottom: 16px;
+}
+.list {
+  @include flex-wrap;
+  justify-content: space-between;
+  gap: 16px;
+  padding-bottom: 32px;
+}
