+++ apps/shop/src/features/product/components/index.ts
@@ -5,3 +5,5 @@ export { default as Option } from './Option';
 export { ProductListLoading } from './ProductListLoading';
 export { default as Description } from './Description';
 export { default as OptionList } from './OptionList';
+export { ProductSection } from './ProductSection';
+export { BannerSlide } from './BannerSlide';
